+++ src/generic/message/mod.rs
@@ -1,0 +1,5 @@
+
+mod message;
+mod json;
+pub use message::Message;
+pub use json::{JSONMessage, Typed};
