+++ src/generic/reactor/reactor.rs
@@ -1,7 +1,28 @@
+use super::*;
+use crate::generic::{
+    BrokerHandle, Handler, LinkOperation, LinkSpawner, Operation, ReactorID, Receiver, Sender,
+};
+
 use std::collections::{HashMap, VecDeque};
 use std::hash::Hash;
 
-use super::*;
+use futures::stream::Stream;
+use futures::task::{Context, Poll};
+use futures::Future;
+use std::pin::Pin;
+
+/// Macro to create reactor handle
+/// This does not borrow the entire Reactor like a function would
+macro_rules! reactorHandle {
+    ($e:expr) => {
+        ReactorHandle::new(
+            &$e.channels.0,
+            &$e.id,
+            &mut $e.inner_ops,
+            &mut $e.broker,
+        );
+    };
+}
 
 /// Gives the option for an init function on a reactor
 pub trait ReactorState<K, M> {
@@ -10,19 +31,6 @@ pub trait ReactorState<K, M> {
 
 /// Blanket implementation for ()
 impl<K, M> ReactorState<K, M> for () {}
-
-/// Macro to create reactor handle
-/// This does not borrow the entire Reactor like a function would
-macro_rules! reactorHandle {
-    ($e:expr) => {
-        ReactorHandle {
-            chan: &$e.channels.0,
-            id: &$e.id,
-            inner_ops: &mut $e.inner_ops,
-            broker: &mut $e.broker,
-        };
-    };
-}
 
 ///
 /// Reactor is the meat and the potatoes of MOZAIC
@@ -71,11 +79,12 @@ where
         params: CoreParams<S, K, M>,
         channels: (Sender<K, M>, Receiver<K, M>),
     ) -> Self {
+        let (state, msg_handlers) = params.consume();
         Reactor {
             id,
             broker,
-            state: params.state,
-            msg_handlers: params.handlers,
+            state,
+            msg_handlers,
             links: HashMap::new(),
             channels,
             inner_ops: VecDeque::new(),
@@ -210,10 +219,6 @@ where
     }
 }
 
-use futures::stream::Stream;
-use futures::task::{Context, Poll};
-use std::pin::Pin;
-
 /// Reactors get spawned with tokio, they only read from their channel and act on the messages
 /// They reduce over an OperationStream
 impl<S, K, M> Future for Reactor<S, K, M>
@@ -261,112 +266,5 @@ where
 {
     fn drop(&mut self) {
         println!("Dropping {:?}", self.id);
-    }
-}
-
-/// Inner op for reactors
-pub enum InnerOp<K, M> {
-    OpenLink(ReactorID, LinkSpawner<K, M>, bool),
-    CloseLink(ReactorID),
-    Close(),
-}
-
-/// Handle to the reactor, managing operation and messages
-pub struct ReactorHandle<'a, K, M> {
-    chan: &'a Sender<K, M>,
-    id: &'a ReactorID,
-    inner_ops: &'a mut VecDeque<InnerOp<K, M>>,
-    broker: &'a mut BrokerHandle<K, M>,
-}
-
-impl<'a, K, M> ReactorHandle<'a, K, M>
-where
-    K: 'static + Send + Eq + Hash + Unpin,
-    M: 'static + Send,
-{
-    pub fn open_link<L>(&mut self, target: ReactorID, spawner: L, cascade: bool)
-    where
-        L: Into<LinkSpawner<K, M>>,
-    {
-        self.inner_ops
-            .push_back(InnerOp::OpenLink(target, spawner.into(), cascade));
-    }
-
-    pub fn open_reactor_like(&mut self, target: ReactorID, tx: Sender<K, M>) {
-        self.broker.spawn_reactorlike(target, tx);
-    }
-
-    pub fn close(&mut self) {
-        self.inner_ops.push_back(InnerOp::Close());
-    }
-
-    pub fn spawn<S: 'static + Send + ReactorState<K, M> + Unpin>(
-        &mut self,
-        params: CoreParams<S, K, M>,
-        id: Option<ReactorID>,
-    ) -> ReactorID {
-        self.broker.spawn(params, id)
-    }
-
-    pub fn id(&mut self) -> &'a ReactorID {
-        &self.id
-    }
-
-    pub fn chan(&self) -> Sender<K, M> {
-        self.chan.clone()
-    }
-}
-
-#[derive(Eq, PartialEq, Debug)]
-pub enum TargetReactor {
-    All,
-    Reactor,
-    Links,
-    Link(ReactorID),
-}
-// ANCHOR Implementation with any::TypeId
-/// Generic implementation of reactor handle, this one is able to handle every T
-/// Making it generic by forming a Message and sending it through
-///
-/// You would want to implement this again with Capnproto messages
-/// to be able to send them over the internet
-impl<'a, K, M> ReactorHandle<'a, K, M> {
-    pub fn send_internal<T: 'static + IntoMessage<K, M>>(&mut self, msg: T, to: TargetReactor) {
-        if let Some((id, msg)) = T::into_msg(msg) {
-            self.chan
-                .unbounded_send(Operation::InternalMessage(id, msg, to))
-                .expect("crashed");
-        }
-    }
-}
-
-// ANCHOR Params
-/// Builder pattern for constructing reactors
-pub struct CoreParams<S, K, M> {
-    state: S,
-    handlers:
-        HashMap<K, Box<dyn for<'a> Handler<S, ReactorHandle<'a, K, M>, (&'a K, &'a mut M)> + Send>>,
-}
-
-impl<S, K, M> CoreParams<S, K, M>
-where
-    K: Eq + Hash + 'static,
-    M: 'static,
-{
-    pub fn new(state: S) -> Self {
-        CoreParams {
-            state,
-            handlers: HashMap::new(),
-        }
-    }
-
-    pub fn handler<H, J>(mut self, handler: H) -> Self
-    where
-        H: Into<(K, J)>,
-        J: for<'a> Handler<S, ReactorHandle<'a, K, M>, (&'a K, &'a mut M)> + Send + 'static,
-    {
-        let (id, handler) = handler.into();
-        self.handlers.insert(id, Box::new(handler));
-        self
     }
 }
